pyproject.toml
[project]
name = "antimony-bootstrap"
version = "0.1.0"
description = "Extract disease mechanisms from knowledge sources and produce parameterized Antimony ODE models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "antimony>=2.15",
    "python-libsbml>=5.20",
    "tellurium>=2.2",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "httpx>=0.27",
    "rich>=13.0",
]

[project.scripts]
antimony-bootstrap = "antimony_bootstrap.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/antimony_bootstrap"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
